requirements.txt
black
ruff
mypy
tox
psycopg
python-dotenv
Flask
flask-sqlalchemy
gunicorn
structlog
pytest
pytest-randomly
pytest-cov
